Moves components into organized folders
