+++ components/Home/BottomAnimation/BottomAnimation.module.scss
