+++ components/Home/BottomAnimation/index.tsx
