+++ components/Home/Partners/index.tsx
