+++ components/Home/Partners/partners.module.scss
