+++ components/deprecated/ShopCategories/index.tsx
