+++ components/deprecated/SuggestedItems/suggestedItems.module.scss
