+++ components/ecommerce/IndividualItem/index.tsx
