+++ components/ecommerce/IndividualItem/individualItem.module.scss
