+++ components/shared/Header/header.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/variables' as vars;
+@use 'styles/variables' as vars;
 $breakpoint: 1100px;
 
 .headerParent {
