+++ components/shared/Loading/Spinner/Spinner.module.scss
