+++ pages/contact/index.tsx
@@ -1,4 +1,4 @@
-import Layout from "components/Layout";
+import Layout from "components/shared/Layout";
 import { NextPage } from "next";
 import ContactForm from "components/ContactForm";
 const contact: NextPage = () => {
